work in progress refactoring. application not working
